+++ src/core/components/io_array.rs
@@ -19,7 +19,7 @@ impl<T: Clone+Default> OutputArray<T> {
     /// Creates a new output array with `num_channels` channels.
     ///
     /// These channels are initialized to time `t=0`.
-    pub fn new(num_channels: uint) -> OutputArray<T> {
+    pub fn new(num_channels: usize) -> OutputArray<T> {
         let mut chs = Vec::with_capacity(num_channels);
         for _ in range(0, num_channels) {
             chs.push(ChannelRef::new());
@@ -28,24 +28,24 @@ impl<T: Clone+Default> OutputArray<T> {
     }
     
     /// Returns the number of output channels
-    pub fn get_num_channels(&self) -> uint {
+    pub fn get_num_channels(&self) -> usize {
         self.chs.len()
     }
 
     /// Returns a reference to channel `i`.
-    pub fn get_channel(&self, i: uint) -> ChannelRef<T> {
+    pub fn get_channel(&self, i: usize) -> ChannelRef<T> {
         self.chs[i].clone()
     }
 
     /// Attempts to get the data frame for time `t` in channel `i`.
     #[inline]
-    pub fn get(&self, i: uint, t: Time) -> Option<T> {
+    pub fn get(&self, i: usize, t: Time) -> Option<T> {
         self.chs[i].get(t)
     }
 
     /// Pushes the next data frame to channel `i`.
     #[inline]
-    pub fn push(&self, i: uint, f: T) {
+    pub fn push(&self, i: usize, f: T) {
         self.chs[i].push(f);
     }
 }
@@ -62,7 +62,7 @@ impl<T: Clone+Default> InputArray<T> {
     /// 
     /// These channels are initialized as empty, and must be filled to return
     /// input data.
-    pub fn new(num_channels: uint) -> InputArray<T> {
+    pub fn new(num_channels: usize) -> InputArray<T> {
         let mut chs = Vec::with_capacity(num_channels);
         for _ in range(0, num_channels) {
             chs.push(None);
@@ -71,17 +71,17 @@ impl<T: Clone+Default> InputArray<T> {
     }
 
     /// Returns the number of input channels
-    pub fn get_num_channels(&self) -> uint {
+    pub fn get_num_channels(&self) -> usize {
         self.chs.len()
     }
 
     /// Sets channel `i` to read from `channel`.
-    pub fn set_channel(&mut self, i: uint, channel: ChannelRef<T>) {
+    pub fn set_channel(&mut self, i: usize, channel: ChannelRef<T>) {
         self.chs[i] = Some(channel);
     }
 
     /// Sets channel `i` to empty.
-    pub fn remove_channel(&mut self, i: uint) {
+    pub fn remove_channel(&mut self, i: usize) {
         self.chs[i] = None;
     }
 
@@ -90,7 +90,7 @@ impl<T: Clone+Default> InputArray<T> {
     /// Returns `None` if either channel `i` is empty, or if channel `i` could
     /// not return the sample for time `t`.
     #[inline]
-    pub fn get(&self, i: uint, t: Time) -> Option<T> {
+    pub fn get(&self, i: usize, t: Time) -> Option<T> {
         match self.chs[i] {
             Some(ref ch) => ch.get(t),
             None => None
