+++ src/dynamics/compressor.rs
@@ -18,7 +18,7 @@ pub struct Compressor {
     pub outputs: OutputArray<Sample>,
 
     level_detectors: Vec<LevelDetector>,
-    num_channels: uint, 
+    num_channels: usize, 
     threshold: f32,
     compression_ratio: f32,
     gain: f32
@@ -34,7 +34,7 @@ impl Compressor {
     /// * The specified `gain` (in decibels) will be applied to the
     ///   signal after compression.
     pub fn new(threshold: f32, compression_ratio: f32, gain: f32, 
-               num_channels: uint) -> Compressor {
+               num_channels: usize) -> Compressor {
         // Create our level detectors
         let mut levels = Vec::with_capacity(num_channels);
         for _ in range(0, num_channels) {
